Refactored class loops and if statements

+++ app/Caching/Cache.php
@@ -5,6 +5,7 @@ namespace App\Caching;
 use App\Interfaces\ICache;
 use App\Logger\Logger;
 use App\Logger\LogLevel;
+use http\Exception\InvalidArgumentException;
 use Memcached;
 
 class Cache implements ICache
@@ -21,6 +22,10 @@ class Cache implements ICache
 
     public function get($key, $default = null) : mixed
     {
+        if (!isset($key)) {
+            throw new InvalidArgumentException('Cache key must not be null.');
+        }
+
         if ($this->memcached->get($key) === false)
         {
             $this->logger->log(LogLevel::CRITICAL, "Cache key '{$key}' does not exist.");
@@ -32,11 +37,19 @@ class Cache implements ICache
 
     public function set($key, $value, $ttl = self::DEFAULT_TTL_SECONDS) : void
     {
+        if (!isset($key)) {
+            throw new InvalidArgumentException("Key is null when setting to cache.");
+        }
+
         $this->memcached->set($key, $value, $ttl);
     }
 
     public function delete($key) : void
     {
+        if (!isset($key)) {
+            throw new InvalidArgumentException('Cache key must not be null.');
+        }
+
         $this->memcached->delete($key);
     }
 
@@ -47,6 +60,10 @@ class Cache implements ICache
 
     public function getMultiple(array $keys, $default = null)
     {
+        if (!is_iterable($keys)) {
+            throw new InvalidArgumentException('Cache keys must be iterable.');
+        }
+
         if ($this->memcached->getMulti($keys) === false)
         {
             $this->logger->log(LogLevel::CRITICAL, "Cache method 'GetMulti' failed.");
@@ -58,19 +75,31 @@ class Cache implements ICache
 
     public function setMultiple(array $values, $ttl = self::DEFAULT_TTL_SECONDS) : void
     {
+        if (!is_iterable($values)) {
+            throw new InvalidArgumentException('Cache values must be iterable.');
+        }
+
         $this->memcached->setMulti($values, $ttl);
     }
 
     public function deleteMultiple(array $keys) : void
     {
+        if (!is_iterable($keys)) {
+            throw new InvalidArgumentException('Cache keys must be iterable.');
+        }
+
         $this->memcached->deleteMulti($keys);
     }
 
     public function has($key) : bool
     {
+        if (!isset($key)) {
+            throw new InvalidArgumentException('Cache key must not be null.');
+        }
+
         if ($this->memcached->get($key) === false)
         {
-            $this->logger->log(LogLevel::CRITICAL, "Cache does not have the key '{$key}'.");
+            $this->logger->log(LogLevel::ALERT, "Cache does not have the key '{$key}'.");
             return false;
         }
